fix: remove circular dependency on getChainflipApi
