+++ bouncer/.eslintrc.json
@@ -50,6 +50,7 @@
     "import/extensions": "off",
     "import/prefer-default-export": "off",
     "import/no-unresolved": "off",
+    "import/named": "error",
     "no-await-in-loop": "off",
     "no-console": "off",
     "no-plusplus": "off",
